added size selection when adding to cart

+++ src/context/CartContext.js
@@ -2,11 +2,13 @@ import { createContext, useState, useEffect } from 'react'
 import { productsArray } from '../data/productData.js'
 
 class CartProduct{
-    constructor(id){
+    constructor(id, size){
         this.id = id; 
         this.name = this.getName(id);
+        this.size = size; 
         this.quantity = 1;
         this.price = this.getPrice(id);
+        this.type = this.getType(id); 
     }
 
     getName(id){
@@ -19,6 +21,12 @@ class CartProduct{
         const product = productsArray.find(product => product.id === id); // find product with id 
 
         return product ? product.price : 0;
+    }
+
+    getType(id){
+        const product = productsArray.find(product => product.id === id); // find product with id
+
+        return product ? product.properties : 0; // get type of product if product was found
     }
 }
 
@@ -39,12 +47,12 @@ export function CartProvider({children}){
         localStorage.setItem('cart', JSON.stringify(cartProducts)) 
     }, [cartProducts])
 
-    function addOneToCart(id){
+    function addItemToCart(id, size){
         const initialCartProducts = cartProducts || []; // initialize cart products if null
 
         // check if product exists 
         for(let i = 0; i < initialCartProducts.length; i++){
-            if(initialCartProducts[i].id === id){
+            if(initialCartProducts[i].id === id && initialCartProducts[i].size === size){
                 const updatedProducts = [...initialCartProducts];
                 updatedProducts[i].quantity += 1; 
                 setCartProducts(updatedProducts);
@@ -52,13 +60,29 @@ export function CartProvider({children}){
             }
         }
 
-        setCartProducts([...initialCartProducts, new CartProduct(id)]); // add product to cart if it does not exist 
+        setCartProducts([...initialCartProducts, new CartProduct(id, size)]); // add product to cart if it does not exist 
     }
 
-    function removeOneFromCart(id) {
+    function addOneToCart(id, size){
+        const initialCartProducts = cartProducts || []; // initialize cart products if null
+
+        // check if product exists 
+        for(let i = 0; i < initialCartProducts.length; i++){
+            if(initialCartProducts[i].id === id && initialCartProducts[i].size === size){
+                const updatedProducts = [...initialCartProducts];
+                updatedProducts[i].quantity += 1; 
+                setCartProducts(updatedProducts);
+                return; 
+            }
+        }
+
+        setCartProducts([...initialCartProducts, new CartProduct(id, size)]); // add product to cart if it does not exist 
+    }
+
+    function removeOneFromCart(id, size) {
         // find product to delete 
         for(let i = 0; i < cartProducts.length; i++){
-            if(cartProducts[i].id === id){
+            if(cartProducts[i].id === id && cartProducts[i].size === size){
                 // remove item from cart
                 if(cartProducts[i].quantity === 1){
                     deleteFromCart(id); 
@@ -72,8 +96,8 @@ export function CartProvider({children}){
         }
     }
 
-    function deleteFromCart(id) {
-        const updatedCartProducts = cartProducts.filter(product => product.id !== id); 
+    function deleteFromCart(id, size) {
+        const updatedCartProducts = cartProducts.filter(product => product.id !== id && product.size !== size); 
 
         setCartProducts(updatedCartProducts)
     }
@@ -91,6 +115,7 @@ export function CartProvider({children}){
 
     const contextValue = {
         items: cartProducts,
+        addItemToCart,
         addOneToCart,
         removeOneFromCart,
         deleteFromCart,
